ck_steps.py
Created ck_steps.py and fixed some issues
