+++ features/Catalina_Gherkin/login_empty_fields.feature
@@ -1,0 +1,6 @@
+# Created by catalinab at 11/8/24
+Feature: # Enter feature name here
+  # Enter feature description here
+
+  Scenario: # Enter scenario name here
+    # Enter steps here
